requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<6.0
psycopg[binary]>=3.2,<3.3
python-dotenv>=1.0,<2.0
django-filter>=24.3,<25.0
drf-spectacular>=0.27,<0.28
pytest>=8.3,<9.0
pytest-django>=4.9,<5.0
ruff>=0.6,<0.7
gunicorn>=23.0,<24.0
Faker>=28.4,<29.0